requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv>=0.21.0